+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.76.5
 
+- Added HPKE operations to (EC)PublicKey---`seal`---and PrivateKey---`open`.